+++ ospd_openvas/notus/metadata.py
@@ -21,15 +21,17 @@
 
 import logging
 
-import csv
 import ast
 import os
 
 from glob import glob
 from hashlib import sha256
 from pathlib import Path
+from csv import DictReader
+from typing import List, Dict
 
-from ospd_openvas import db, nvticache
+from ospd_openvas.db import MainDB
+from ospd_openvas.nvticache import NVTICache
 from ospd_openvas.errors import OspdOpenvasError
 from ospd_openvas.openvas import Openvas
 
@@ -57,42 +59,43 @@ EXPECTED_FIELD_NAMES_LIST = [
 
 METADATA_DIRECTORY_NAME = "notus_metadata"
 
+# Metadata constant field definitions
+SCRIPT_CATEGORY = "3"  # ACT_GATHER_INFO
+SCRIPT_TIMEOUT = "0"
+SCRIPT_FAMILY = "Notus_LSC_Metadata"
+BIDS = ""
+REQUIRED_KEYS = ""
+MANDATORY_KEYS = ""
+EXCLUDED_KEYS = ""
+REQUIRED_UDP_PORTS = ""
+REQUIRED_PORTS = ""
+DEPENDENCIES = ""
+
 
 class NotusMetadataHandler:
     """Class to perform checksum checks and upload metadata for
     CSV files that were created by the Notus Generator."""
 
-    def __init__(self, metadata_path: str = None):
+    def __init__(self, nvti: NVTICache = None, metadata_path: str = None):
+        self._nvti = nvti
+        self._metadata_path = metadata_path
+        self._openvas_settings_dict = None
 
-        openvas_object = Openvas()
-        self.openvas_settings_dict = openvas_object.get_settings()
-        self.__no_signature_check = self.openvas_settings_dict[
-            "nasl_no_signature_check"
-        ]
+    @property
+    def nvti(self) -> NVTICache:
+        if self._nvti is None:
+            try:
+                maindb = MainDB()
+                self._nvti = NVTICache(maindb)
+            except SystemExit:
+                raise OspdOpenvasError(
+                    "Could not connect to the Redis KB"
+                ) from None
 
-        # Figure out the path to the metadata
-        if not metadata_path:
-            self.__metadata_path = self._get_metadata_path()
-        else:
-            self.__metadata_path = metadata_path
+        return self._nvti
 
-        self.__metadata_relative_path_string = "{}/".format(
-            METADATA_DIRECTORY_NAME
-        )
-
-        # Get a list of all CSV files in that directory with their absolute path
-        self.__csv_abs_filepaths_list = self._get_csv_filepaths()
-
-        # Connect to the Redis KB
-        try:
-            self.__db_ctx = db.OpenvasDB.create_context(1)
-            main_db = db.MainDB()
-            self.__nvti_cache = nvticache.NVTICache(main_db)
-        except SystemExit:
-            # Maybe replace this with just a log message
-            raise Exception("Could not connect to the Redis KB") from None
-
-    def _get_metadata_path(self) -> str:
+    @property
+    def metadata_path(self) -> str:
         """Find out where the CSV files containing the metadata
         are on the file system, depending on whether this machine
         is a GSM or GVM in a development environment.
@@ -101,34 +104,41 @@ class NotusMetadataHandler:
             A full path to the directory that contains all Notus
             metadata.
         """
-        # Openvas is installed and the plugins folder configured.
-        plugins_folder = self.openvas_settings_dict.get("plugins_folder")
-        if plugins_folder:
-            metadata_path = "{}/{}/".format(
-                plugins_folder, METADATA_DIRECTORY_NAME
-            )
-            return metadata_path
+        if self._metadata_path is None:
+            # Openvas is installed and the plugins folder configured.
+            plugins_folder = self.openvas_setting.get("plugins_folder")
+            if plugins_folder:
+                self._metadata_path = (
+                    f'{plugins_folder}/{METADATA_DIRECTORY_NAME}/'
+                )
+                return self._metadata_path
 
-        try:
-            # From the development environment - Not used in production
-            install_prefix = os.environ["INSTALL_PREFIX"]
-        except KeyError:
-            install_prefix = None
+            try:
+                # From the development environment - Not used in production
+                install_prefix = os.environ["INSTALL_PREFIX"]
+            except KeyError:
+                install_prefix = None
 
-        if not install_prefix:
-            # Fall back to the path used in production
-            metadata_path = "/opt/greenbone/feed/plugins/{}/".format(
-                METADATA_DIRECTORY_NAME
-            )
-        else:
-            metadata_path = "{}/var/lib/openvas/plugins/{}/".format(
-                install_prefix, METADATA_DIRECTORY_NAME
-            )
+            if not install_prefix:
+                # Fall back to the path used in production
+                self._metadata_path = (
+                    f'/opt/greenbone/feed/plugins/{METADATA_DIRECTORY_NAME}/'
+                )
+            else:
+                self._metadata_path = f'{install_prefix}/var/lib/openvas/plugins/{METADATA_DIRECTORY_NAME}/'  # pylint: disable=C0301
 
-        return metadata_path
+        return self._metadata_path
 
-    def _get_csv_filepaths(self) -> Path:
-        """Get the absolute file path to all detected CSV files
+    @property
+    def openvas_setting(self):
+        """Set OpenVAS option."""
+        if self._openvas_settings_dict is None:
+            openvas_object = Openvas()
+            self._openvas_settings_dict = openvas_object.get_settings()
+        return self._openvas_settings_dict
+
+    def _get_csv_filepaths(self) -> List[Path]:
+        """Get a list of absolute file paths to all detected CSV files
         in the relevant directory.
 
         Returns:
@@ -136,7 +146,7 @@ class NotusMetadataHandler:
         """
         return [
             Path(csv_file).resolve()
-            for csv_file in glob("{}*.csv".format(self.__metadata_path))
+            for csv_file in glob(f'{self.metadata_path}*.csv')
         ]
 
     def _check_field_names_lsc(self, field_names_list: list) -> bool:
@@ -198,10 +208,10 @@ class NotusMetadataHandler:
             Example: URL:www.example.com, URL:www.example2.com
         """
         formatted_list = list()
-        advisory_xref_string = "URL:{}".format(advisory_xref_string)
+        advisory_xref_string = f'URL:{advisory_xref_string}'
         formatted_list.append(advisory_xref_string)
         for url_string in xrefs_list:
-            url_string = "URL:{}".format(url_string)
+            url_string = f'URL:{url_string}'
             formatted_list.append(url_string)
         return ", ".join(formatted_list)
 
@@ -217,7 +227,9 @@ class NotusMetadataHandler:
             Whether the checksum check was successful or not.
             Also returns true if the checksum check is disabled.
         """
-        if not self.__no_signature_check:
+
+        no_signature_check = self.openvas_setting.get("nasl_no_signature_check")
+        if not no_signature_check:
             with file_abs_path.open("rb") as file_file_bytes:
                 sha256_object = sha256()
                 # Read chunks of 4096 bytes sequentially to avoid
@@ -229,8 +241,8 @@ class NotusMetadataHandler:
                 file_calculated_checksum_string = sha256_object.hexdigest()
                 # Extract the downloaded checksum for this file
                 # from the Redis KB
-                file_downloaded_checksum_string = db.OpenvasDB.get_single_item(
-                    self.__db_ctx, "sha256sums:{}".format(str(file_abs_path))
+                file_downloaded_checksum_string = self.nvti.get_file_checksum(
+                    file_abs_path
                 )
 
                 # Checksum check
@@ -242,6 +254,119 @@ class NotusMetadataHandler:
         # Checksum check was either successful or it was skipped
         return True
 
+    def upload_lsc_from_csv_reader(
+        self,
+        file_name: str,
+        general_metadata_dict: Dict,
+        csv_reader: DictReader,
+    ) -> bool:
+        """For each advisory_dict, write its contents to the
+        Redis KB as metadata.
+
+        Arguments:
+            file_name: CSV file name with metadata to be uploaded
+            general_metadata_dict: General metadata common for all advisories
+                                   in the CSV file.
+            csv_reader: DictReader iterator to access the advisories
+
+        Return True if success, False otherwise
+        """
+
+        loaded = 0
+        total = 0
+        for advisory_dict in csv_reader:
+            # Make sure that no element is missing in the advisory_dict,
+            # else skip that advisory
+            total += 1
+            is_correct = self._check_advisory_dict(advisory_dict)
+            if not is_correct:
+                continue
+            # For each advisory_dict,
+            # write its contents to the Redis KB as metadata.
+            # Create a list with all the metadata. Refer to:
+            # https://github.com/greenbone/ospd-openvas/blob/232d04e72d2af0199d60324e8820d9e73498a831/ospd_openvas/db.py#L39 # pylint: disable=C0321
+            advisory_metadata_list = list()
+            # File name
+            advisory_metadata_list.append(
+                f'{METADATA_DIRECTORY_NAME}/{file_name}'
+            )
+            # Required keys
+            advisory_metadata_list.append(REQUIRED_KEYS)
+            # Mandatory keys
+            advisory_metadata_list.append(MANDATORY_KEYS)
+            # Excluded keys
+            advisory_metadata_list.append(EXCLUDED_KEYS)
+            # Required UDP ports
+            advisory_metadata_list.append(REQUIRED_UDP_PORTS)
+            # Required ports
+            advisory_metadata_list.append(REQUIRED_PORTS)
+            # Dependencies
+            advisory_metadata_list.append(DEPENDENCIES)
+            # Tags
+            tags_string = (
+                "cvss_base_vector={}|last_modification={}|"
+                "creation_date={}|summary={}|vuldetect={}|"
+                "insight={}|affected={}|solution={}|"
+                "solution_type={}|qod_type={}"
+            )
+            tags_string = tags_string.format(
+                advisory_dict["CVSS_BASE_VECTOR"],
+                advisory_dict["LAST_MODIFICATION"],
+                advisory_dict["CREATION_DATE"],
+                advisory_dict["DESCRIPTION"],
+                general_metadata_dict["VULDETECT"],
+                advisory_dict["INSIGHT"],
+                advisory_dict["AFFECTED"],
+                general_metadata_dict["SOLUTION"],
+                general_metadata_dict["SOLUTION_TYPE"],
+                general_metadata_dict["QOD_TYPE"],
+            )
+            advisory_metadata_list.append(tags_string)
+            # CVEs
+            advisory_metadata_list.append(
+                ", ".join(ast.literal_eval(advisory_dict["CVE_LIST"]))
+            )
+
+            advisory_metadata_list.append(BIDS)
+            # XREFS
+            advisory_metadata_list.append(
+                self._format_xrefs(
+                    advisory_dict["ADVISORY_XREF"],
+                    ast.literal_eval(advisory_dict["XREFS"]),
+                )
+            )
+
+            # Script category
+            advisory_metadata_list.append(SCRIPT_CATEGORY)
+            # Script timeout
+            advisory_metadata_list.append(SCRIPT_TIMEOUT)
+            # Script family
+            advisory_metadata_list.append(SCRIPT_FAMILY)
+            # Script Name / Title
+            advisory_metadata_list.append(advisory_dict["TITLE"])
+
+            # Write the metadata list to the respective Redis KB key,
+            # overwriting any existing values
+            oid = advisory_dict["OID"]
+            kb_key_string = f'nvt:{oid}'
+            try:
+                self.nvti.add_vt_to_cache(
+                    vt_id=kb_key_string, vt=advisory_metadata_list
+                )
+            except OspdOpenvasError:
+                logger.warning(
+                    "LSC will not be loaded. The advisory_metadata_"
+                    "list was either not a list or does not include "
+                    "15 entries"
+                )
+                continue
+            loaded += 1
+
+        logger.debug(
+            "Loaded %d/%d advisories from %s", loaded, total, file_name
+        )
+        return loaded == total
+
     def update_metadata(self) -> None:
         """Parse all CSV files that are present in the
         Notus metadata directory, perform a checksum check,
@@ -250,14 +375,16 @@ class NotusMetadataHandler:
         """
 
         logger.debug("Starting the Notus metadata load up")
-        # 1. Read each CSV file
-        for csv_abs_path in self.__csv_abs_filepaths_list:
-            # 2. Check the checksums, unless they have been disabled
+        # Get a list of all CSV files in that directory with their absolute path
+        csv_abs_filepaths_list = self._get_csv_filepaths()
+
+        # Read each CSV file
+        for csv_abs_path in csv_abs_filepaths_list:
+            # Check the checksums, unless they have been disabled
             if not self.is_checksum_correct(csv_abs_path):
                 # Skip this file if the checksum does not match
-                logger.debug("Checksum failed")
+                logger.warning('Checksum for %s failed', csv_abs_path)
                 continue
-
             logger.debug("Checksum check for %s successful", csv_abs_path)
             with csv_abs_path.open("r") as csv_file:
                 # Skip the license header, so the actual content
@@ -268,94 +395,22 @@ class NotusMetadataHandler:
                         general_metadata_dict = ast.literal_eval(line_string)
                         break
                 # Check if the file can be parsed by the CSV module
-                reader = csv.DictReader(csv_file)
+                reader = DictReader(csv_file)
                 # Check if the CSV file has the expected field names,
                 # else skip the file
                 is_correct = self._check_field_names_lsc(reader.fieldnames)
                 if not is_correct:
+                    logger.warning(
+                        'Field names check for %s failed', csv_abs_path
+                    )
                     continue
 
-                for advisory_dict in reader:
-                    # Make sure that no element is missing in the advisory_dict,
-                    # else skip that advisory
-                    is_correct = self._check_advisory_dict(advisory_dict)
-                    if not is_correct:
-                        continue
+                file_name = csv_abs_path.name
+                if not self.upload_lsc_from_csv_reader(
+                    file_name, general_metadata_dict, reader
+                ):
+                    logger.warning(
+                        "Some advaisory was not loaded from %s", file_name
+                    )
 
-                    # 3. For each advisory_dict,
-                    # write its contents to the Redis KB as metadata.
-                    # Create a list with all the metadata. Refer to:
-                    # https://github.com/greenbone/ospd-openvas/blob/232d04e72d2af0199d60324e8820d9e73498a831/ospd_openvas/db.py#L39
-                    advisory_metadata_list = list()
-                    # File name
-                    advisory_metadata_list.append(
-                        "{0}{1}".format(
-                            self.__metadata_relative_path_string,
-                            os.path.basename(csv_file.name),
-                        )
-                    )
-                    # Required keys
-                    advisory_metadata_list.append("")
-                    # Mandatory keys
-                    advisory_metadata_list.append("")
-                    # Excluded keys
-                    advisory_metadata_list.append("")
-                    # Required UDP ports
-                    advisory_metadata_list.append("")
-                    # Required ports
-                    advisory_metadata_list.append("")
-                    # Dependencies
-                    advisory_metadata_list.append("")
-                    # Tags
-                    tags_string = (
-                        "cvss_base_vector={}|last_modification={}|"
-                        "creation_date={}|summary={}|vuldetect={}|"
-                        "insight={}|affected={}|solution={}|"
-                        "solution_type={}|qod_type={}"
-                    )
-                    tags_string = tags_string.format(
-                        advisory_dict["CVSS_BASE_VECTOR"],
-                        advisory_dict["LAST_MODIFICATION"],
-                        advisory_dict["CREATION_DATE"],
-                        advisory_dict["DESCRIPTION"],
-                        general_metadata_dict["VULDETECT"],
-                        advisory_dict["INSIGHT"],
-                        advisory_dict["AFFECTED"],
-                        general_metadata_dict["SOLUTION"],
-                        general_metadata_dict["SOLUTION_TYPE"],
-                        general_metadata_dict["QOD_TYPE"],
-                    )
-                    advisory_metadata_list.append(tags_string)
-                    # CVEs
-                    advisory_metadata_list.append(
-                        ", ".join(ast.literal_eval(advisory_dict["CVE_LIST"]))
-                    )
-                    # BIDs
-                    advisory_metadata_list.append("")
-                    # XREFS
-                    advisory_metadata_list.append(
-                        self._format_xrefs(
-                            advisory_dict["ADVISORY_XREF"],
-                            ast.literal_eval(advisory_dict["XREFS"]),
-                        )
-                    )
-                    # Script Category
-                    advisory_metadata_list.append("3")
-                    # Timeout
-                    advisory_metadata_list.append("0")
-                    # Script Family
-                    advisory_metadata_list.append("Notus_LSC_Metadata")
-                    # Script Name / Title
-                    advisory_metadata_list.append(advisory_dict["TITLE"])
-
-                    # Write the metadata list to the respective Redis KB key,
-                    # overwriting any existing values
-                    kb_key_string = "nvt:{}".format(advisory_dict["OID"])
-                    try:
-                        self.__nvti_cache.add_vt_to_cache(
-                            vt_id=kb_key_string, vt=advisory_metadata_list
-                        )
-                    except OspdOpenvasError:
-                        # The advisory_metadata_list was either not
-                        # a list or does not include 15 entries
-                        break
+        logger.debug("Notus metadata load up finished.")